requirements.txt
aiogram==3.*
SQLAlchemy==2.*
aiosqlite
python-dotenv
